update: refactor the sharewb dialog to remove mui and use our components
