+++ webapp/app.ironcalc.com/frontend/src/components/ShareWorkbook/share-workbook-modal.css
@@ -1,0 +1,80 @@
+.share-modal-backdrop {
+  background-color: transparent;
+}
+
+@media (max-width: 600px) {
+  .share-modal-backdrop {
+    background-color: rgb(0 0 0 / 32%);
+  }
+}
+
+.share-modal-paper {
+  position: absolute;
+  top: 44px;
+  right: 0px;
+  margin: 10px;
+  width: 340px;
+  max-width: calc(100% - 20px);
+}
+
+.share-modal-content {
+  flex-direction: row;
+  gap: 12px;
+}
+
+.share-modal-url-wrapper {
+  display: flex;
+  flex-direction: column;
+  min-height: 100%;
+  width: 100%;
+  justify-content: space-between;
+}
+
+.share-modal-qr-wrapper {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.share-modal-qr-wrapper svg {
+  width: 72px;
+  height: auto;
+}
+
+@media (max-width: 600px) {
+  .share-modal-paper {
+    position: static;
+    margin: 0;
+  }
+
+  .share-modal-content {
+    flex-direction: column;
+    align-items: center;
+  }
+
+  .share-modal-qr-wrapper svg {
+    width: 160px;
+  }
+
+  .share-modal-url-wrapper {
+    min-height: unset;
+  }
+
+  .share-modal-footer {
+    flex-direction: row;
+    align-items: center;
+  }
+}
+
+.share-modal-footer {
+  justify-content: flex-start;
+  border-top: 1px solid var(--palette-grey-300);
+  font-size: var(--typography-font-size);
+  font-family: var(--typography-font-family);
+  color: var(--palette-grey-600);
+}
+
+.share-modal-footer svg {
+  width: 16px;
+  flex-shrink: 0;
+}
