--- webapp/app.ironcalc.com/frontend/src/components/ShareWorkbookDialog.tsx
@@ -1,212 +1,0 @@
-import type { Model } from "@ironcalc/workbook";
-import { Button, Dialog, styled, TextField } from "@mui/material";
-import { Check, Copy, GlobeLock } from "lucide-react";
-import { QRCodeSVG } from "qrcode.react";
-import { useEffect, useState } from "react";
-import { useTranslation } from "react-i18next";
-import { shareModel } from "./rpc";
-
-function ShareWorkbookDialog(properties: {
-  onClose: () => void;
-  onModelUpload: (blob: ArrayBuffer, fileName: string) => Promise<void>;
-  model?: Model;
-}) {
-  const [url, setUrl] = useState<string>("");
-  const [copied, setCopied] = useState(false);
-  const { t } = useTranslation();
-
-  useEffect(() => {
-    const generateUrl = async () => {
-      if (properties.model) {
-        const bytes = properties.model.toBytes();
-        const hash = await shareModel(bytes);
-        setUrl(`${location.origin}/?model=${hash}`);
-      }
-    };
-    generateUrl();
-  }, [properties.model]);
-
-  useEffect(() => {
-    let timeoutId: ReturnType<typeof setTimeout>;
-    if (copied) {
-      timeoutId = setTimeout(() => {
-        setCopied(false);
-      }, 2000);
-    }
-    return () => {
-      if (timeoutId) {
-        clearTimeout(timeoutId);
-      }
-    };
-  }, [copied]);
-
-  const handleClose = () => {
-    properties.onClose();
-  };
-
-  const handleCopy = async () => {
-    try {
-      await navigator.clipboard.writeText(url);
-      setCopied(true);
-    } catch (err) {
-      console.error("Failed to copy text: ", err);
-    }
-  };
-
-  return (
-    <DialogWrapper
-      open={true}
-      tabIndex={0}
-      onClose={handleClose}
-      transitionDuration={0}
-      onKeyDown={(event) => {
-        if (event.code === "Escape") {
-          handleClose();
-        }
-      }}
-    >
-      <DialogContent>
-        <QRCodeWrapper>
-          <QRCodeSVG value={url} />
-        </QRCodeWrapper>
-        <URLWrapper>
-          <StyledTextField
-            hiddenLabel
-            disabled
-            value={url}
-            variant="outlined"
-            fullWidth
-            margin="normal"
-            size="small"
-          />
-          <StyledButton
-            variant="contained"
-            color="primary"
-            size="small"
-            onClick={handleCopy}
-          >
-            {copied ? <StyledCheck /> : <StyledCopy />}
-            {copied
-              ? t("file_bar.share_popover.copied")
-              : t("file_bar.share_popover.copy_url")}
-          </StyledButton>
-        </URLWrapper>
-      </DialogContent>
-
-      <UploadFooter>
-        <GlobeLock />
-        {t("file_bar.share_popover.info_text")}
-      </UploadFooter>
-    </DialogWrapper>
-  );
-}
-
-const DialogWrapper = styled(Dialog)`
-  .MuiDialog-paper {
-    width: 340px;
-    position: absolute;
-    top: 44px;
-    right: 0px;
-    margin: 10px;
-    max-width: calc(100% - 20px);
-    border-radius: 8px;
-    box-shadow: 0px 4px 12px rgba(0, 0, 0, 0.15);
-  }
-  .MuiBackdrop-root {
-    background-color: transparent;
-  }
-`;
-
-const DialogContent = styled("div")`
-  padding: 16px;
-  display: flex;
-  flex-direction: row;
-  gap: 12px;
-  height: auto;
-`;
-
-const URLWrapper = styled("div")`
-  display: flex;
-  flex-direction: column;
-  min-height: 100%;
-  width: 100%;
-  justify-content: space-between;
-`;
-
-const StyledTextField = styled(TextField)`
-  margin: 0px;
-  .MuiInputBase-root {
-    max-height: 36px;
-    font-size: 14px;
-    padding-top: 0px;
-  }
-  .MuiOutlinedInput-input {
-    text-overflow: ellipsis;
-    padding: 8px;
-  }
-`;
-
-const StyledButton = styled(Button)`
-  display: flex;
-  flex-direction: row;
-  gap: 4px;
-  background-color: #eeeeee;
-  height: 36px;
-  color: #616161;
-  box-shadow: none;
-  font-size: 12px;
-  text-transform: capitalize;
-  gap: 10px;
-  &:hover {
-    background-color: #e0e0e0;
-    box-shadow: none;
-  }
-  &:active {
-    background-color: #d4d4d4;
-    box-shadow: none;
-  }
-`;
-
-const StyledCopy = styled(Copy)`
-  width: 16px;
-`;
-
-const StyledCheck = styled(Check)`
-  width: 16px;
-`;
-
-const QRCodeWrapper = styled("div")`
-  height: auto;
-  border-radius: 4px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  svg {
-    width: 80px;
-    height: auto;
-  }
-
-  @media (max-width: 600px) {
-    display: none;
-  }
-`;
-
-const UploadFooter = styled("div")`
-  height: 44px;
-  border-top: 1px solid #e0e0e0;
-  font-size: 12px;
-  font-weight: 400;
-  color: #757575;
-  display: flex;
-  align-items: center;
-  font-family: Inter;
-  gap: 8px;
-  padding: 0px 12px;
-  svg {
-    max-width: 16px;
-    min-width: 16px;
-  }
-`;
-
-export default ShareWorkbookDialog;
